requirements.txt
flask
Flask-Cors 
requests
psycopg2
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
werkzeug
python-dotenv
PyJWT
Flask-JWT-Extended